Feat(kotlin): add bunch of new text objects
The list of supported text objects for the language Kotlin were
absolutely minimal so far. To improve the language support, this adds
the most common objects.
But neither are all possible text objects supported yet, nor are all
possible trees covered. Though it should be a fairly good start to work
with.