fix: properly handle ranges ending with a newline (#828)
Fixes #827 . The changes are similar to the ones in #812 , but they are
different from the function mentioned in a comment of that PR
(https://github.com/nvim-treesitter/nvim-treesitter-textobjects/pull/812#issuecomment-3479015036)
because said function assumes that the range returned is end-inclusive,
while this plugin currently is treating the range as end_exclusive.

This bug was surfaced by #687 (because and `end_row, end_col` pair like
`4, 0` becomes `4, -1` after the change), but this wasn't working
properly before #687 either. The behavior before that PR was to also
select the first character of the next line (so, one character after the
newline).