fix(repeatable_move): using `;` or `,` in operator-pending mode misses one character. ci: fix broken test (#833)
Fixes #699 properly.

Also fixes broken tests.

I personally found it very hard to test the CI, because it's hard to
test locally due to the fact that all parsers are installed into
`~/.local/share/nvim`. Thus, I think it's good to sandbox the data into
`.test-deps` in `minimal_init.lua`.

Also I believe that lint-docs should be a separate CI that opens a PR of
a new updated docs.